requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
psycopg2-binary
Shapely
geojson
pyproj
python-jose
pyjwt
passlib
python-multipart